requirements.txt
Django>=4.0
channels>=4.0.0
django-crispy-forms>=2.0
django-crispy-bootstrap5>=2023.0
channels-redis>=4.1.0
daphne>=4.0.0
redis>=5.0.0
PyMySQL>=1.1.0
mysqlclient>=2.2.0